backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.10
fastapi-cli==0.0.7
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.15
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
setuptools==69.0.3
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlmodel==0.0.23
starlette==0.46.0
typer==0.15.2
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
